ctl: rename element event mask
The 'snd_ctl_event' structure can be defined to deliver several types of
events. One of the event is for element and it includes bitflags of the
kind of event for the element. In this library, the bitflag is named so
that it represents generic event, however it should be something
specific to element event.

This commit renames the bitflag.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>