Add support workflow of Github Actions
Github Actions is tool to automate software development workflows for CI/CD.
This tool has a remarkable point to share 'action' in repositories in github,
which is a unit of processing defined by users. The users write and push
'workflow' with the 'action' into repository in github.com. When any workflow
file is added to 'default' branch, the automation runs[2].

The defined actions run on virtual machine in Microsoft Azure (Standard_DS2_v2
as of today[2]). Three types of operating system are available: Windows, MacOS
and Linux (Ubuntu). On the MacOS and Linux machine, 'sudo' is available to
setup host environment for CI/CD, as well as Docker container is supported on
the host.

This commit adds a workflow file to automate build test. In this automation,
Fedora 32 in Docker container and Ubuntu 19.10 in LXD container as build
environments in Linux host. The automation is triggered in each 'push' and
'pull_request' event. The result is available in 'Actions' tab of repository
page in github.com.

[1] https://help.github.com/en/actions/automating-your-workflow-with-github-actions/about-github-actions
[2] https://help.github.com/en/actions/automating-your-workflow-with-github-actions/configuring-a-workflow
[3] https://help.github.com/en/actions/automating-your-workflow-with-github-actions/virtual-environments-for-github-hosted-runners

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>