timer: user-instance: delegate decision of poll timeout into application
As long as using the created Gsource with GMainContext, the call of poll(2)
system call with infinite timeout surely returns when quit() method is
called for the context. All of GMainContext implements GWakeup with eventfd
and the call of quit() emits event via file descriptor of eventfd. This
brings wakeup from blocking when poll(2) is called with inifinite timeout.

This means that it's necessarily need to have explicit timeout internally.
This commit removes implementation of GSourceFunc.prepare for the reason.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>