Revert "seq: tstamp: use wrapper structure instead of union"
This reverts commit 20fdc2d893ccba609cc786ebac733bf6914f9874.

In the above commit, union type structure is wrapped with care of
language bindings in which union-compatible feature is not expected.
However, it's responsible for language bindings to parse gir and
the care is not necessarily required in shared library side. In first
place, ALSASeq.Tstamp is boxed type object.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>