timer: user_instance: add API to choose the type of event to read
Currently ALSATimer.UserInstance has any attach API to decide the type of
event to read. In design of ALSA Timer core, the attach API has two
functionality; to attach or to detach with error for absent timer device.
Current implementation has inconvenient for the functionality due to
execution to choose the type of event in advance. The execution fails
when the instance is already attached.

This commit adds a new API to choose the type of event and split the
call from any attach API.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>