seq: query: use GFileError to report error
In design of ALSA sequencer, sequencer character device is not
corresponding to each substance in ALSA Sequencer. It's a kind of
interface to query and operate. Applications adds user client
and operates it to use its services. Nevertheless, the interface
has some functionality without adding the user instances.

For the functionality, this library has global functions. The functions
are not relevant to the user instance, therefore it's not reasonable to
use error domain specific to the user instance.

This commit uses GFileError to report errors in the global functions.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>