rawmidi: report open/read/write/fcntl error
The call of open/read/write system call can return several types of error
from VFS. It's hard to dispatch all of the error in local error domain.

This commit uses GFileError to dispatch the most of the error. At
failure of conversion to GFileError, local error domain is used.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>