ctl: card: change prototype of function relevant to TLV
In ALSA control interface, the data of TLV (Type-Length-Value) style is
defined as array with unsigned int type of element. However, current
implementation of ALSACtl.Card handles them as signed int type.

This commit fixes the bug. The commit loses backward compatibility to
v0.1.0 release in a point of type of function argument. The exported
symbols relevant to TLV are updated with new version.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>