seq: user-client: use QueueTimerCommon interface instead of QueueTimer
Current implementation to handle queue timer is by ALSASeq.QueueTimer
gobject class, and uses boxed object (=structure) to express actual data
of the timer. The implementation is not necessarily convenient to add
more type of queue timer since it requires to add getter/setter of data
to the class.

This commit uses ALSASeq.QueueTimerCommon interface to handle any type
of queue timer in ALSASeq.UserInstance for system call.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>