seq: event-cntr/event-data-queue: obsolete timestamp functions by adding alternatives
In UAPI of ALSA Sequencer, time stamp is defined as union which has
tick time member and real time member. There is no tag to select two
members, thus user applications should judge which member is used. In
this meaning, exposing the union is not necessarily useful.

This commit is a preparation to obsolete the union. The methods to
retrieve the union are rewrite to retrieve either tick time or real time
explicitly.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>