event: add public API to compute pool consumption when scheduling event
In design of ALSA Sequencer, delivery of event consumes cell in client
pool. For usual event, it's one cell. For variable length of event, it's
multiple cells including the total length of event.

ALSASeq.ClientPool class has properties to express the number of cells
and free cells in client pool. When comparing to the properties,
applications get to know whether the scheduling operation is blocked or
not.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>