all: remove annotation from error domain
As I fixed gi-docgen at 6c690567b6b ("utils: fix failure to generate
link to error domain"), gi-docgen release of 2022.2 brings loss of
backward compatibility about annotation for error domain.

This commit is a workaround for successful build either in 2022.1 or
former releases and in 2022.2 or later releases of gi-docgen. When
gi-docgen 2022.2 disseminates vastly, the annotation will be back.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>