timer: optimize to gi-docgen 2024.1 or later, approximately
As long as building with gi-docgen 2025.3.1 in Ubuntu 25.04, some
annotations for SlaveClass become invalid and causes some warnings.

/usr/bin/gi-docgen generate --no-namespace-dir --config=doc/timer/alsatimer.toml --output-dir=doc/timer/alsatimer --content-dir=/home/mocchi/git/alsa-project.org/alsa-gobject/doc/timer src/timer/ALSATimer-0.0.gir
WARNING: Unknown namespace SlaveClass
slave_id. If the slave_class is [enum@SlaveClass.APPLICATION], the slave_id is for the PID of
                                ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~
WARNING: Unknown namespace SlaveClass
[enum@SlaveClass.SEQUENCER], the slave_id is the numeric ID of queue bound for timer device.
^~~~~~~~~~~~~~~~~~~~~~~~~~~
WARNING: Unknown namespace SlaveClass
slave_id. If the slave_class is [enum@SlaveClass.APPLICATION], the slave_id is for the PID of
                                ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~
WARNING: Unknown namespace SlaveClass
[enum@SlaveClass.SEQUENCER], the slave_id is the numeric ID of queue bound for timer device.
^~~~~~~~~~~~~~~~~~~~~~~~~~~
WARNING: Unknown namespace SlaveClass
slave_id. If the slave_class is [enum@SlaveClass.APPLICATION], the slave_id is for the PID of
                                ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~
WARNING: Unknown namespace SlaveClass
[enum@SlaveClass.SEQUENCER], the slave_id is the numeric ID of queue bound for timer device.
^~~~~~~~~~~~~~~~~~~~~~~~~~~

This commit suppresses the warnings by using the pattern of enum@class.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>